pyproject.toml
[project]
name = "spider-xhs"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "loguru>=0.7.3",
    "openpyxl>=3.1.5",
    "pyexecjs>=1.5.1",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "retry>=0.9.2",
]
